requirements.txt
openai
python-dotenv
pydantic
fastapi
uvicorn
boto3
gunicorn